+++ public/app/features/alerting/unified/components/RuleConditionSection.tsx
@@ -7,7 +7,6 @@ import { Trans, t } from '@grafana/i18n';
 import {
   Combobox,
   type ComboboxOption,
-  Field,
   Icon,
   InlineField,
   InlineFieldRow,
@@ -22,12 +21,11 @@ import { ToLabel } from 'app/features/expressions/components/ToLabel';
 import { ExpressionDatasourceUID, reducerTypes, thresholdFunctions } from 'app/features/expressions/types';
 import { isRangeEvaluator } from 'app/features/expressions/utils/expressionTypes';
 
-import { evaluateEveryValidationOptions } from '../group-details/validation';
 import { createSimpleConditionExpressions } from '../rule-editor/formProcessing';
 import { type RuleFormValues, type SimpleCondition } from '../types/rule-form';
 
 import { EvaluationGroupFieldRow } from './rule-editor/EvaluationGroupFieldRow';
-import { EvaluationGroupQuickPick } from './rule-editor/EvaluationGroupQuickPick';
+import { RuleEvaluationIntervalField } from './rule-editor/RuleEvaluationIntervalField';
 
 const DEFAULT_SIMPLE_CONDITION: SimpleCondition = {
   whenField: ReducerID.last,
@@ -43,15 +41,8 @@ interface RuleConditionSectionProps {
 
 export function RuleConditionSection({ hideEvaluationGroup = false }: RuleConditionSectionProps = {}) {
   const base = useStyles2(getStyles);
-  const {
-    watch,
-    setValue,
-    getValues,
-    register,
-    formState: { errors },
-  } = useFormContext<RuleFormValues>();
+  const { watch, setValue, getValues } = useFormContext<RuleFormValues>();
   const evaluateFor = watch('evaluateFor') || '0s';
-  const evaluateEvery = watch('evaluateEvery');
 
   const [simpleCondition, setSimpleCondition] = useState<SimpleCondition>(DEFAULT_SIMPLE_CONDITION);
 
@@ -190,25 +181,7 @@ export function RuleConditionSection({ hideEvaluationGroup = false }: RuleCondit
           </InlineFieldRow>
 
           {hideEvaluationGroup ? (
-            <Stack direction="column" gap={1.5}>
-              <Field
-                noMargin
-                label={t('alerting.rule-form.evaluation.interval-label', 'Evaluation interval')}
-                error={errors.evaluateEvery?.message}
-                invalid={Boolean(errors.evaluateEvery?.message)}
-                htmlFor="evaluate-every-no-group"
-              >
-                <Input
-                  id="evaluate-every-no-group"
-                  width={8}
-                  {...register('evaluateEvery', evaluateEveryValidationOptions<{ evaluateEvery: string }>([]))}
-                />
-              </Field>
-              <EvaluationGroupQuickPick
-                currentInterval={evaluateEvery}
-                onSelect={(interval) => setValue('evaluateEvery', interval)}
-              />
-            </Stack>
+            <RuleEvaluationIntervalField />
           ) : (
             <EvaluationGroupFieldRow enableProvisionedGroups={false} />
           )}
